# apps/learninggoalhub/client/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/courses/{courseId}/extract": {
      "post": {
        "tags": [
          "extraction-controller"
        ],
        "operationId": "extract",
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "model",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionStartResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/api/extractions/current": {
      "get": {
        "tags": [
          "current-extraction-controller"
        ],
        "operationId": "current",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentExtractionResponse"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/courses/{courseId}/competency-tree": {
      "post": {
        "tags": [
          "extraction-controller"
        ],
        "operationId": "rebuildCompetencyTree",
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "model",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CompetencyTreeResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/{id}": {
      "get": {
        "tags": [
          "course-controller"
        ],
        "operationId": "getCourse",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CourseSummaryResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "course-controller"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "course-controller"
        ],
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/courses/{courseId}/extract/status": {
      "get": {
        "tags": [
          "extraction-controller"
        ],
        "operationId": "extractStatus",
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Snapshot"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CourseResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "outputLanguage": {
            "type": "string"
          },
          "figuresEnabled": {
            "type": "boolean"
          }
        }
      },
...
      "ExtractionSummary": {
        "type": "object",
        "properties": {
          "documentsProcessed": {
            "type": "integer",
            "format": "int32"
          },
          "goalsCreated": {
            "type": "integer",
            "format": "int32"
          },
          "terminalCompetencies": {
            "type": "integer",
            "format": "int32"
          },
          "textSources": {
            "type": "integer",
            "format": "int32"
          },
          "figureSources": {
            "type": "integer",
            "format": "int32"
          },
          "unsupportedSources": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ExtractionStartResponse": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string",
            "enum": [
              "RUNNING",
              "SUCCEEDED",
              "FAILED"
            ]
          }
        }
      },
      "CurrentExtractionResponse": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "integer",
            "format": "int64"
          },
          "courseName": {
            "type": "string"
          },
          "phase": {
            "type": "string",
            "enum": [
              "DESCRIBING_FIGURES",
              "OUTLINING",
              "PARSING",
              "EXTRACTING",
              "CLASSIFYING",
              "EMBEDDING",
              "PERSISTING",
              "SYNTHESIZING"
            ]
          },
          "percent": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "enum": [
              "RUNNING",
              "SUCCEEDED",
              "FAILED"
            ]
          }
        }
      },
...
      "CompetencyTreeResult": {
        "type": "object",
        "properties": {
          "competencies": {
            "type": "integer",
            "format": "int32"
          },
          "unmatchedGoals": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "UpdateCourseRequest": {
        "type": "object",
        "properties": {
          "outputLanguage": {
            "type": "string"
          },
          "figuresEnabled": {
            "type": "boolean"
          }
        }
      },
...
      "CourseSummaryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "outputLanguage": {
            "type": "string"
          },
          "figuresEnabled": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "documentCount": {
            "type": "integer",
            "format": "int64"
          },
          "goalCount": {
            "type": "integer",
            "format": "int64"
          },
          "skillCount": {
            "type": "integer",
            "format": "int64"
          },
          "skillsReviewedAt": {
            "type": "string",
            "format": "date-time"
          },
          "extractionStatus": {
            "type": "string",
            "enum": [
              "RUNNING",
              "SUCCEEDED",
              "FAILED"
            ]
          }
        }
      },
...
      "Snapshot": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "RUNNING",
              "SUCCEEDED",
              "FAILED"
            ]
          },
          "phase": {
            "type": "string",
            "enum": [
              "DESCRIBING_FIGURES",
              "OUTLINING",
              "PARSING",
              "EXTRACTING",
              "CLASSIFYING",
              "EMBEDDING",
              "PERSISTING",
              "SYNTHESIZING"
            ]
          },
          "completed": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "percent": {
            "type": "integer",
            "format": "int32"
          },
          "model": {
            "type": "string"
          },
          "summary": {
            "$ref": "#/components/schemas/ExtractionSummary"
          },
          "error": {
            "type": "string"
          },
          "failedSessions": {
            "type": "integer",
            "format": "int32"
          },
          "failedSessionNames": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}